+++ silo-core/contracts/utils/ShareDebtToken.sol
@@ -20,12 +20,6 @@ contract ShareDebtToken is IERC20R, ShareToken {
     /// @dev maps _owner => _recipient => amount
     mapping(address => mapping(address => uint256)) private _receiveAllowances;
 
-    error OwnerIsZero();
-    error RecipientIsZero();
-    error ShareTransferNotAllowed();
-    error AmountExceedsAllowance();
-    error RecipientNotSolventAfterTransfer();
-
     /// @param _silo Silo address for which tokens was deployed
     function initialize(ISilo _silo, address _hookReceiver) external virtual initializer {
         __ShareToken_init(_silo, _hookReceiver);
@@ -103,8 +97,12 @@ contract ShareDebtToken is IERC20R, ShareToken {
         // if we are minting or burning, Silo is responsible to check all necessary conditions
         // if we are NOT minting and not burning, it means we are transferring
         // make sure that _recipient is solvent after transfer
-        if (_isTransfer(_sender, _recipient) && !silo.isSolvent(_recipient)) {
-            revert RecipientNotSolventAfterTransfer();
+        if (_isTransfer(_sender, _recipient)) {
+            ISilo cacheSilo =  silo;
+            // reading from silo directly is a bit more gas friendly, than creating storage for it
+            (address silo0, address silo1) = cacheSilo.config().getSilos();
+            ISilo collateralSilo = ISilo(address(cacheSilo) == silo0 ? silo1 : silo0);
+            if (!collateralSilo.isSolvent(_recipient)) revert RecipientNotSolventAfterTransfer();
         }
     }
 }